Rename timer.c and timer.h to timerfd
