+++ src/main/host/host.c
@@ -30,7 +30,7 @@
 #include "main/host/descriptor/regular_file.h"
 #include "main/host/descriptor/socket.h"
 #include "main/host/descriptor/tcp.h"
-#include "main/host/descriptor/timer.h"
+#include "main/host/descriptor/timerfd.h"
 #include "main/host/descriptor/transport.h"
 #include "main/host/descriptor/udp.h"
 #include "main/host/futex_table.h"
