+++ src/main/host/syscall/protected.c
@@ -15,7 +15,7 @@
 #include "lib/logger/logger.h"
 #include "main/host/descriptor/descriptor.h"
 #include "main/host/descriptor/tcp.h"
-#include "main/host/descriptor/timer.h"
+#include "main/host/descriptor/timerfd.h"
 #include "main/host/syscall_condition.h"
 #include "main/host/thread.h"
